Do not prefill input with the last history entry
We used to show the last entry from the history *if* `filters` was empty.
This was a fast way to switch the last entry on and off, basically the user
committed an empty string to "reset" the filter, and then `f` plus `enter`
to get the previous filter back.  But now we have a better `F` for just
that.

It is slightly better to not prefill because the automatic reading of such
a value that always occurs does disturb the thinking of the user.